+++ examples/lib/examples_web/router.ex
@@ -9,6 +9,6 @@ defmodule ExamplesWeb.Router do
     pipe_through(:api)
 
     resources("/users", UserController, only: [:create, :index, :show, :delete])
-    resources("/orders", OrderController, only: [:create, :index, :show, :delete])
+    resources("/orders", OrderController, only: [:create, :index, :show, :delete, :update])
   end
 end